fix #350: hydration issue with SSR/SSG and re-declaration of TableContext (#351)
* typo(table): fix re-declaration of TableContext

TableContext is already defined as the type and also redeclared a React.Context. Hence renamed the typedef as TableContextType and update the Table component

Signed-off-by: chankruze <chankruze@gmail.com>

* fix(lib): proper window check in ssr/ssg pages

While rendering Modal, there was a difference between the React tree that was pre-rendered (SSR/SSG) and the React tree that rendered during the first render in the Browser (Hydration). Due to which there was issues with hydration when used in SSR/SSG. This issue was caused by the windowExists() that was using window in a component's rendering.

Signed-off-by: chankruze <chankruze@gmail.com>

* fix(lib): add missing dependency in Modal useEffect

Signed-off-by: chankruze <chankruze@gmail.com>

Signed-off-by: chankruze <chankruze@gmail.com>